Don't interfere with `setuptools.sh`'s shebang
`python-libs/setuptools.sh` must run inside a `bash` (as it
uses `pushd`/`popd`), and its shebang says so. Therefore, the
Makefile must not launch it as `sh setuptools.sh`, because
that overrides the shebang, and thus the desired shell,
to `sh`.

This commit changes `setuptools.sh`'s file access mode to
allow the Makefile to directly launch it, `./setuptools.sh`.
It also changes the Makefile accordingly.